requirements.txt
matplotlib>=3.9.0
numpy>=1.26.4
pillow>=10.3.0
networkx>=3.3
Flask>=3.0.3
imageio>=2.34.2